+++ src/v1.x/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,59 @@
 # ui
 
+## 1.51.0
+
+### Patch Changes
+
+- 4b28c78: testing changeset versioning
+- 2839a15: Update versioning strategy
+- Updated dependencies [2839a15]
+- Updated dependencies [2afd4e3]
+  - @copilotkit/runtime-client-gql@1.51.0
+  - @copilotkit/react-core@1.51.0
+  - @copilotkit/shared@1.51.0
+
+## 1.51.0-next.4
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.51.0-next.4
+- @copilotkit/react-core@1.51.0-next.4
+- @copilotkit/shared@1.51.0-next.4
+
+## 1.51.0-next.3
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.51.0-next.3
+- @copilotkit/react-core@1.51.0-next.3
+- @copilotkit/shared@1.51.0-next.3
+
+## 1.51.0-next.2
+
+### Patch Changes
+
+- 4b28c78: testing changeset versioning
+- Updated dependencies [2afd4e3]
+  - @copilotkit/shared@1.51.0-next.2
+  - @copilotkit/react-core@1.51.0-next.2
+  - @copilotkit/runtime-client-gql@1.51.0-next.2
+
+## 1.51.0-next.1
+
+### Patch Changes
+
+- @copilotkit/react-core@1.51.0-next.1
+- @copilotkit/runtime-client-gql@1.51.0-next.1
+- @copilotkit/shared@1.51.0-next.1
+
+## 1.50.2-next.0
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.50.2-next.0
+- @copilotkit/react-core@1.50.2-next.0
+- @copilotkit/shared@1.50.2-next.0
+
 ## 1.50.1
 
 ### Patch Changes
@@ -2678,7 +2732,6 @@
 - e16d95e: New prerelease
 - 1721cbd: Add convertActionsToDynamicStructuredTools to sdk-js
 - CopilotKit Core:
-
   - Improved error messages and overall logs
   - `useCopilotAction.renderAndAwait` renamed to `.renderAndAwaitForResponse` (backwards compatible, will be deprecated in the future)
   - Improved scrolling behavior. It is now possible to scroll up during LLM response generation
@@ -2686,7 +2739,6 @@
   - Updated interfaces for better developer ergonomics
 
   CoAgents:
-
   - Renamed `remoteActions` to `remoteEndpoints` (backwards compatible, will be deprecated in the future)
   - Support for LangGraph Platform in Remote Endpoints
   - LangGraph JS Support for CoAgents (locally via `langgraph dev`, `langgraph up` or deployed to LangGraph Platform)
@@ -2694,7 +2746,6 @@
   - Enhanced state management and message handling
 
   CopilotKid Back-end SDK:
-
   - Released a whole-new `@copilotkit/sdk-js` for building agents with LangGraph JS Support
 
 - 8d0144f: bump
@@ -3097,7 +3148,6 @@ CopilotKid Back-end SDK:
 ### Patch Changes
 
 - inject minified css in bundle
-
   - removes the need to import `styles.css` manually
   - empty `styles.css` included in the build for backwards compatibility
   - uses tsup's `injectStyles` with `postcss` to bundle and minify the CSS, then inject it as a style tag
@@ -3262,7 +3312,6 @@ CopilotKid Back-end SDK:
 ### Major Changes
 
 - b6a4b6eb: V1.0 Release Candidate
-
   - A robust new protocol between the frontend and the Copilot Runtime
   - Support for Copilot Cloud
   - Generative UI
@@ -3270,7 +3319,6 @@ CopilotKid Back-end SDK:
   - OpenAI assistant API streaming
 
 - V1.0 Release
-
   - A robust new protocol between the frontend and the Copilot Runtime
   - Support for Copilot Cloud
   - Generative UI
@@ -3314,7 +3362,6 @@ CopilotKid Back-end SDK:
 ### Major Changes
 
 - V1.0 Release Candidate
-
   - A robust new protocol between the frontend and the Copilot Runtime
   - Support for Copilot Cloud
   - Generative UI