+++ src/v1.x/packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @copilotkit/sdk-js
 
+## 1.51.0
+
+### Minor Changes
+
+- ce14f8b: create copilotkit langgraph middleware
+
+### Patch Changes
+
+- 2839a15: Update versioning strategy
+- Updated dependencies [2839a15]
+- Updated dependencies [2afd4e3]
+  - @copilotkit/shared@1.51.0
+
+## 1.51.0-next.4
+
+### Patch Changes
+
+- @copilotkit/shared@1.51.0-next.4
+
+## 1.51.0-next.3
+
+### Patch Changes
+
+- @copilotkit/shared@1.51.0-next.3
+
+## 1.51.0-next.2
+
+### Patch Changes
+
+- Updated dependencies [2afd4e3]
+  - @copilotkit/shared@1.51.0-next.2
+
+## 1.51.0-next.1
+
+### Minor Changes
+
+- ce14f8b: create copilotkit langgraph middleware
+
+### Patch Changes
+
+- @copilotkit/shared@1.51.0-next.1
+
+## 1.50.2-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.50.2-next.0
+
 ## 1.50.1
 
 ### Patch Changes
@@ -1890,7 +1938,6 @@
 - e16d95e: New prerelease
 - 1721cbd: Add convertActionsToDynamicStructuredTools to sdk-js
 - CopilotKit Core:
-
   - Improved error messages and overall logs
   - `useCopilotAction.renderAndAwait` renamed to `.renderAndAwaitForResponse` (backwards compatible, will be deprecated in the future)
   - Improved scrolling behavior. It is now possible to scroll up during LLM response generation
@@ -1898,7 +1945,6 @@
   - Updated interfaces for better developer ergonomics
 
   CoAgents:
-
   - Renamed `remoteActions` to `remoteEndpoints` (backwards compatible, will be deprecated in the future)
   - Support for LangGraph Platform in Remote Endpoints
   - LangGraph JS Support for CoAgents (locally via `langgraph dev`, `langgraph up` or deployed to LangGraph Platform)
@@ -1906,7 +1952,6 @@
   - Enhanced state management and message handling
 
   CopilotKid Back-end SDK:
-
   - Released a whole-new `@copilotkit/sdk-js` for building agents with LangGraph JS Support
 
 - 8d0144f: bump